[tts-jump-back] fix broken tests
